+++ example1/android/app/src/main/kotlin/com/example/example1/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.example1
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}