+++ example_getx/lib/src/controller/controller1.dart
@@ -1,0 +1,11 @@
+import 'package:get/get.dart';
+
+class GetxController1 extends GetxController {
+  static GetxController1 get to => Get.find();
+  var list = <String>[
+    'abc',
+    'dda',
+    'cvzc',
+    'ggfg',
+  ].obs;
+}