+++ example_getx/lib/src/controller/controller2.dart
@@ -1,0 +1,5 @@
+import 'package:get/get.dart';
+
+class GetxController2 extends GetxController {
+  RxInt count = 0.obs;
+}